finish test cleanup
