+++ airbyte-integrations/connectors/source-notion/unit_tests/test_components.py
@@ -4,16 +4,6 @@
 
 import pytest
 from source_notion.components import *
-from source_notion.source import SourceNotion
-
-
-def test_streams():
-    source = SourceNotion()
-    config_mock = {"start_date": "2020-01-01T00:00:00.000Z",
-                   "credentials": {"auth_type": "token", "token": "abcd"}}
-    streams = source.streams(config_mock)
-    expected_streams_number = 5
-    assert len(streams) == expected_streams_number
 
 
 def test_notion_properties_transformation():
